borme_flow/requirements.txt
pymysql
requests
pdfplumber
beautifulsoup4
cryptography
python-dotenv
sentry-sdk
